+++ NEWS.md
@@ -8,6 +8,7 @@
 
 * Add a limited-scope `exclude` argument to `tar_watch()` and `tar_watch_server()` (#458, @gorkang).
 * Write a `.gitignore` file to ignore everything in `_targets/meta/` except `.gitignore` and `_targets/meta/meta`.
+* Target Markdown: add `knitr` engines for pipeline construction and prototyping from within literate programming documents (#469, @nviets, @emilyriederer, @cderv, @ijlyttle, @GShotwell, @gadenbuie, @tomsing1).
 
 ## Enhancements